refactor(praxis-detail): convert to hook + dispatcher + Default archetype
Splits the ~600-line PraxisDetail monolith into usePraxisDetail (the stable
PraxisDetailState contract: praxis+votes fetch, withdraw/resubmit/flag/moderate
handlers, admin-bar visibility, isOwner) + a thin dispatcher (keyed on
praxis.task_faction_slug via pickVariant, owning the loading/error/not-found
guards) + DefaultPraxisDetail (the old layout lifted verbatim). No faction has a
bespoke archetype yet, so every faction renders the Default path unchanged.
Voting still routes through VoteUI, so faction vote UIs keep working.
